FAVICON via Server-Side Includes
Various web browsers and mobile platforms want a web site icon in lots
of sizes and formats.

An online tool at https://realfavicongenerator.net creates them all,
along with platform configuration files.

The HTML pointing to the favicon will be the same for every page, so I
broke it out into a text fragment that can be folded in via Server-Side
Include.